% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{getCuratedBreastDataExprSetList}
\alias{getCuratedBreastDataExprSetList}
\title{getCuratedBreastDataExprSetList}
\usage{
getCuratedBreastDataExprSetList(test = FALSE)
}
\arguments{
\item{test}{A boolean. If TRUE, downloads a miniature test dataset instead of the full dataset. Default is FALSE.}
}
\value{
A list, with each index containing an ExpressionSet object from a specific study, and potentially a specific batch.
}
\description{
A list of ExpressionSet objects, one for each curated study, containing
study-specific gene expression and phenotype data. FeatureNames are gene
symbols. Data is already quantile normalized according to standard protocols
for 1 and 2-channel arrays, depending on the platform used for this study.
}
\details{
Batches from studies are treated as individual datasets, as the signal
can differ between batches. Thus, an expression object named using a GSE
study number followed by an underscore means this ExpressionSet contains
samples either from a distinct platform (and the study used >1 platforms),
or from a distinct batch or tissue site.  An "all" tag means that there were no
batches for this study. Raw data files downloaded from GEO oftentimes have clear
batch/site information appended to sample names; this was often the source of
batch identification and how the package developer chose to create the batch
name string.
}
\examples{
curatedBreastDataExprSetList <- getCuratedBreastDataExprSetList(test = TRUE)
# what are all the names of the studies?
names(curatedBreastDataExprSetList)
# what is the dimension of the gene expression matrix for study GSE2034?
dim(exprs(curatedBreastDataExprSetList$study_2034_GPL96_all))
}
\references{
Planey, Butte. Database integration of 4923 publicly-available samples of breast cancer molecular and clinical data. AMIA Joint Summits Translational Science Proceedings. (2003) PMC3814460
}
