% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedBreastData_process.R
\name{collapseDupProbes}
\alias{collapseDupProbes}
\title{Collapse/handle duplicated probes (genes) in a dataset}
\usage{
collapseDupProbes(
  expr,
  sampleColNames = colnames(expr),
  keys,
  method = c("average", "highestVariance"),
  debug = TRUE,
  removeNA_keys = TRUE,
  varMetric = c("everything", "all.obs", "complete.obs", "na.or.complete",
    "pairwise.complete.obs")
)
}
\arguments{
\item{expr}{An expression matrix with genes in the rows and samples in the columns.}

\item{sampleColNames}{Sample column names. Needed for internal debugging; usually the default colnames(expr) is appropriate.}

\item{keys}{Generally the list of gene symbols, or some molecular key, that needs to be "collapsed" because it contains duplicated names.}

\item{method}{Method used to collapse probes: take the mean across all duplicated keys, or just pick the key with the highest variance?}

\item{debug}{Use internal unit tests that will stop the code if it detects a bug?}

\item{removeNA_keys}{Remove any NA keys?}

\item{varMetric}{Standard options taken from the base var() function. May be important if you have NA values in your data matrix; otherwise, "everything" is usually fine.}
}
\value{
Returns a processed list with the items "expr" and "keys", the expression matrix and final keys list.
}
\description{
Used internally by processExpressionSet.  Code to either take the average across
a set of duplicated "keys" (can be probes or genes, which correspond to the rows
in the expression matrix "expr"), or take the keys that has the highest variance 
across the set of duplicated keys.
}
\examples{
#load up our datasets
curatedBreastDataExprSetList <- getCuratedBreastDataExprSetList(test=TRUE)

#just perform on second dataset, GSE2034, as an example.
#This dataset has no NAs already but does have duplicated genes
#highestVariance calculation make take a minute to run.
collapsedData <- collapseDupProbes(expr=exprs(curatedBreastDataExprSetList[[2]]),  
keys=fData(curatedBreastDataExprSetList[[2]])[, "gene_symbol"], 
method = c("highestVariance"), debug = TRUE, removeNA_keys = TRUE, 
varMetric = c("everything"))
#look at names of outputs
names(collapsedData)
}
\author{
Katie Planey <katie.planey@gmail.com>
}
