% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{curatedBreastData-package}
\alias{curatedBreastData}
\alias{curatedBreastData-package}
\title{Curated breast gene expression data with survival and treatment information}
\description{
34 manually curated high-quality gene expression microarray datasets with 
advanced breast cancer samples collected from GEO. All datasets provided have 
some form of survival and treatment information, and all such clinical variables
are semantically normalized across all datasets for easy analyses across 
datasets. Authors of the PubMed article linked to each GEO dataset were
contacted in an effort to collect as much extra clinical data as possible.
See vignette and publication reference from AMIA Translational Science Joint
Summits presentation in 2013 for more details on how this data was curated.
}
\details{
Functions are provided to post-process standard S4 ExpressionSet objects 
to remove samples with high NA rates, impute missing values, collapse duplicated
gene symbols or probes, remove duplicated samples that share the same patient
ID, and filter genes by variance magnitude or percentile.
}
\author{
Katie Planey
}
\keyword{internal}
