% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curatedBreastData_pData}
\alias{curatedBreastData_pData}
\title{curatedBreastData_pData}
\format{
A \code{data.frame} with patient samples as rows and phenotype variables as columns.
}
\usage{
data(curatedBreastData_pData)
}
\description{
A single \code{data.frame} containing the concatenated \code{pData} from all datasets in the \code{curatedBreastDataExprSetList} list. Rownames correspond to sample IDs.
}
\details{
This data object was created by combining the \code{pData} slot of all ExpressionSets provided in the \code{curatedBreastDataExprSetList}.
}
\note{
Suggestions for new datasets to add are always welcome; the maintainer does aim
to only include datasets that have minimal treatment and some form of survival
(and/or treatment response) to allow for richer analyses. Raw data is always
preferred in order to control normalization schemes. Normalization details for
each dataset can be found in the GitHub repo in the References section.
}
\examples{
#load up datasets that are in S4 expressionSet format.
curatedBreastDataExprSetList <- getCuratedBreastDataExprSetList(test=TRUE)

#process only the first two datasets to avoid a long-running example:
#take top 100 genes by variance from each dataset.
proc_curatedBreastDataExprSetList <- processExpressionSetList(exprSetList=
curatedBreastDataExprSetList[1:2], 
outputFileDirectory = tempdir(), numTopVarGenes=100)

#now we have processed expression matrices,
#each with the top 100 genes by variance 
data(curatedBreastData_pData)
dim(curatedBreastData_pData)
head(curatedBreastData_pData)
}
\references{
Planey, Butte. Database integration of 4923 publicly-available samples of breast
cancer molecular and clinical data. AMIA Joint Summits Translational
Science Proceedings. (2003) PMC3814460

GitHub repo with code, further documentation on datasets and baseline
normalization schemes, and database quality checks: 
https://github.com/kplaney/curatedBreastCancer
}
\keyword{datasets}
\keyword{package}
